Update filter components
- Fix rendering issue
- Correctly compare and update timeRange
- Split out files for simplicity
